smp: Move more common code out of platform
